pyproject.toml
[project]
name = "hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "google-cloud-speech>=2.38.0",
    "google-genai>=1.68.0",
    "pyaudio>=0.2.14",
    "pydantic-settings>=2.13.1",
    "sounddevice>=0.5.0",
    "numpy>=2.0.0",
    "lightrag-hku>=1.0.0",
    "loguru>=0.7.3",
    "textual>=8.2.0",
]


[tool.ruff]
unsafe-fixes = true
exclude = [".venv"]
extend-include = ["*.ipynb"]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
preview = true
explicit-preview-rules = true
select = [
    # Default rules
    "E",     # pycodestyle errors
    "F",     # Pyflakes
    "C4",    # flake8-comprehensions
    "C90",   # mccabe complex structure
    "D",     # pydocstyle
    "I",     # isort
    "PT",    # flake8-pytest-style
    "PL",    # Pylint
    "SIM",   # flake8-simplify
    "UP",    # pyupgrade
    "W",     # pycodestyle warnings
    "S",     # flake8-bandit
    "ANN",   # flake8-annotations
    "B",     # flake8-bugbear
    "NPY",   # NumPy-specific rules
    "UP045",
]
extend-select = [
    "RUF022",  # unsorted-dunder-all
    "PLR6301", # no-self-use
    "PLC0415",
    "E402",    # module-import-not-at-top-of-file
]
ignore = [
    "B024",
    "B028",    # no-explicit-stacklevel
    "B904",
    "C901",    # complex-structure
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D107",
    "D415",
    "D200",
    "D205",
    "D212",
    "E722",
    "ANN002",
    "ANN003",
    "ANN204",  # dynamically typed expressions
    "ANN401",
    "PLR0913",
    "PLR2004", # magic value used in comparison
    "PT028",   # test function parameter has default argument
    "PLR6301", # could be a static method
    "NPY002",
]

[tool.ruff.lint.pylint]
max-locals = 20

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 120

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]
